Add display data table
1. Create data-table component
2. Create extrac-data service to get data from model

Notice the click on the aside item takes 350ms for scripting to
extra the data. Implement service to cache data for each item,
I guess model hook in details is expensive too. may need to
optimize in future.

Possible solution for optimize:

- Solution 1: Put details page as a component of /profile instead of
a subroute, then we can pass data down to the route.

- Solution 2: Implement a service for the current model hook of
detail page, cache the data in service.